clean codebase

+++ src/Vulkan/Instance.cpp
@@ -1,7 +1,6 @@
 #include <iostream>
 #include "../Oreginum/Core.hpp"
 #include "../Oreginum/Window.hpp"
-#include "../Oreginum/Logger.hpp"
 #include "Instance.hpp"
 
 VKAPI_ATTR VkResult VKAPI_CALL vkCreateDebugReportCallbackEXT(VkInstance instance,
@@ -31,59 +30,28 @@ VKAPI_ATTR VkBool32 VKAPI_CALL debug_callback_function(VkDebugReportFlagsEXT fla
 
 Oreginum::Vulkan::Instance::Instance(bool debug)
 {
-	Logger::info("Creating Vulkan instance with debug: " + std::string(debug ? "enabled" : "disabled"), true);
-	
 	vk::ApplicationInfo application_information
 	{Oreginum::Window::get_title().c_str(), VK_MAKE_VERSION(0, 0, 1), "Oreginum Engine",
 		VK_MAKE_VERSION(0, 0, 1), VK_API_VERSION_1_0};
 
-	// Log base extensions
-	Logger::info("Base Vulkan extensions: VK_KHR_SURFACE_EXTENSION_NAME, VK_KHR_WIN32_SURFACE_EXTENSION_NAME");
-
-	if(debug) {
-		instance_extensions.push_back(VK_EXT_DEBUG_REPORT_EXTENSION_NAME);
+	if(debug) instance_extensions.push_back(VK_EXT_DEBUG_REPORT_EXTENSION_NAME),
 		instance_layers.push_back("VK_LAYER_LUNARG_standard_validation");
-		Logger::info("Added debug extension: VK_EXT_DEBUG_REPORT_EXTENSION_NAME");
-		Logger::info("Added validation layer: VK_LAYER_LUNARG_standard_validation");
-	}
-
-	Logger::info("Total extensions: " + std::to_string(instance_extensions.size()) + ", layers: " + std::to_string(instance_layers.size()));
 
 	vk::InstanceCreateInfo instance_information{{}, &application_information,
 		static_cast<uint32_t>(instance_layers.size()), instance_layers.data(),
 		static_cast<uint32_t>(instance_extensions.size()), instance_extensions.data()};
 
-	vk::Result result = vk::createInstance(&instance_information, nullptr, instance.get());
-	if(result != vk::Result::eSuccess) {
-		Logger::excep("Failed to create Vulkan instance: VkResult " + std::to_string(static_cast<int>(result)));
-		Oreginum::Core::error("Vulkan is not supported sufficiently.");
-	}
-
-	Logger::info("Vulkan instance created successfully with " + std::to_string(instance_extensions.size()) + " extensions", true);
+	if(vk::createInstance(&instance_information, nullptr, instance.get()) !=
+		vk::Result::eSuccess) Oreginum::Core::error("Vulkan is not supported sufficiently.");
 
 	if(debug) create_debug_callback();
 }
 
 Oreginum::Vulkan::Instance::~Instance()
 {
-	if(instance.use_count() != 1) {
-		Logger::info("Vulkan instance destructor: shared instance, skipping cleanup");
-		return;
-	}
-	
-	Logger::info("Destroying Vulkan instance", true);
-	
-	if(debug_callback) {
-		Logger::info("Destroying debug report callback");
-		instance->destroyDebugReportCallbackEXT(debug_callback, nullptr);
-	}
-	
-	if(*instance) {
-		Logger::info("Destroying Vulkan instance handle");
-		instance->destroy();
-	}
-	
-	Logger::info("Vulkan instance cleanup completed");
+	if(instance.use_count() != 1) return;
+	if(debug_callback) instance->destroyDebugReportCallbackEXT(debug_callback, nullptr);
+	if(*instance) instance->destroy();
 }
 
 void Oreginum::Vulkan::Instance::swap(Instance *other)
@@ -96,22 +64,12 @@ void Oreginum::Vulkan::Instance::swap(Instance *other)
 
 void Oreginum::Vulkan::Instance::create_debug_callback()
 {
-	Logger::info("Creating Vulkan debug report callback");
-	
 	vk::DebugReportCallbackCreateInfoEXT debug_callback_information;
-	debug_callback_information.flags = vk::DebugReportFlagBitsEXT::eError |
+	debug_callback_information.flags = vk::DebugReportFlagBitsEXT::eError | 
 		vk::DebugReportFlagBitsEXT::eWarning | vk::DebugReportFlagBitsEXT::ePerformanceWarning;
 	debug_callback_information.pfnCallback = reinterpret_cast<vk::PFN_DebugReportCallbackEXT>(debug_callback_function);
 
-	Logger::info("Debug callback flags: Error | Warning | PerformanceWarning");
-
-	vk::Result result = instance->createDebugReportCallbackEXT(&debug_callback_information,
-		nullptr, &debug_callback);
-	
-	if(result != vk::Result::eSuccess) {
-		Logger::excep("Failed to create debug report callback: VkResult " + std::to_string(static_cast<int>(result)));
+	if(instance->createDebugReportCallbackEXT(&debug_callback_information,
+		nullptr, &debug_callback) != vk::Result::eSuccess)
 		Oreginum::Core::error("Could not initialize Vulkan debugging.");
-	}
-	
-	Logger::info("Vulkan debug report callback created successfully");
 }
